+++ android/src/main/java/ee/forgr/capacitor_updater/CapacitorUpdater.java
@@ -488,12 +488,15 @@ public class CapacitorUpdater {
   }
 
   private String getChecksum(File file) throws IOException {
-    byte[] bytes = new byte[(int) file.length()];
-    try (FileInputStream fis = new FileInputStream(file)) {
-      fis.read(bytes);
-    }
+    final int BUFFER_SIZE = 1024 * 1024 * 5; // 5 MB buffer size
     CRC32 crc = new CRC32();
-    crc.update(bytes);
+    try (FileInputStream fis = new FileInputStream(file)) {
+      byte[] buffer = new byte[BUFFER_SIZE];
+      int length;
+      while ((length = fis.read(buffer)) != -1) {
+        crc.update(buffer, 0, length);
+      }
+    }
     String enc = String.format("%08X", crc.getValue());
     return enc.toLowerCase();
   }