+++ ios/Plugin/CapacitorUpdater.swift
@@ -332,9 +332,27 @@ extension CustomError: LocalizedError {
     }
 
     private func getChecksum(filePath: URL) -> String {
+        let bufferSize = 1024 * 1024 * 5 // 5 MB
+        var checksum = uLong(0)
+
         do {
-            let fileData: Data = try Data.init(contentsOf: filePath)
-            let checksum: uLong = fileData.withUnsafeBytes { crc32(0, $0.bindMemory(to: Bytef.self).baseAddress, uInt(fileData.count)) }
+            let fileHandle = try FileHandle(forReadingFrom: filePath)
+            defer {
+                fileHandle.closeFile()
+            }
+
+            while autoreleasepool(invoking: {
+                let fileData = fileHandle.readData(ofLength: bufferSize)
+                if fileData.count > 0 {
+                    checksum = fileData.withUnsafeBytes {
+                        crc32(checksum, $0.bindMemory(to: Bytef.self).baseAddress, uInt(fileData.count))
+                    }
+                    return true // Continue
+                } else {
+                    return false // End of file
+                }
+            }) {}
+
             return String(format: "%08X", checksum).lowercased()
         } catch {
             print("\(self.TAG) Cannot get checksum: \(filePath.path)", error)